// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CollegeStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type CollegeStatus = (typeof CollegeStatus)[keyof typeof CollegeStatus]


export const SectionType = {
  HERO: 'HERO',
  ABOUT: 'ABOUT',
  COURSES: 'COURSES',
  ADMISSIONS: 'ADMISSIONS',
  FACULTY: 'FACULTY',
  FACILITIES: 'FACILITIES',
  GALLERY: 'GALLERY',
  EVENTS: 'EVENTS',
  CONTACT: 'CONTACT',
  FOOTER: 'FOOTER',
  PLACEMENTS: 'PLACEMENTS',
  TESTIMONIALS: 'TESTIMONIALS',
  DOWNLOADS: 'DOWNLOADS',
  FEES: 'FEES',
  ACCREDITATION: 'ACCREDITATION',
  HEADER: 'HEADER',
  FEATURES: 'FEATURES',
  CTA: 'CTA',
  CUSTOM: 'CUSTOM'
} as const

export type SectionType = (typeof SectionType)[keyof typeof SectionType]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  DISABLED: 'DISABLED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const AccessRequestStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type AccessRequestStatus = (typeof AccessRequestStatus)[keyof typeof AccessRequestStatus]